re-orgnize file structure
